+++ scripts/dev_setup.py
@@ -11,7 +11,6 @@ import sys
 import os
 from subprocess import check_call, CalledProcessError
 
-modules = ['batch-extensions', 'batch-cli-extensions']
 root_dir = os.path.abspath(os.path.join(os.path.abspath(__file__), '..', '..'))
 
 
@@ -25,6 +24,7 @@ def exec_command(command):
         sys.exit(1)
 
 print('Running dev setup...')
+print(os.environ)
 print('Root directory \'{}\'\n'.format(root_dir))
 
 # install general requirements and azure-cli
@@ -33,8 +33,8 @@ exec_command('pip install -r requirements.txt')
 # install automation package
 exec_command('pip install -e ./scripts')
 
-for m in modules:
-  # install reference to extension module package
-  exec_command('pip install -e {}'.format(os.path.join(root_dir, m)))
+# install reference to extension module package
+exec_command('pip install -e {}'.format(root_dir))
+exec_command('pip install --upgrade --target ./.azure/devcliextensions/azure-batch-cli-extensions {0}/batch-cli-extensions'.format(root_dir))
 
 print('Finished dev setup.')